requirements.txt
click==8.0.3
cv==1.0.0
cycler==0.11.0
fonttools==4.29.1
importlib-metadata==4.11.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
nltk==3.7
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
regex==2022.1.18
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.62.3
typing_extensions==4.1.1
wordcloud==1.8.1
zipp==3.7.0
